Reorganized tests into folders + minor changes to account for concurrency issues
